Remove extra info from H2Core::Note (including get_position(), ADSR, and filter params)
Note had a lot of information in it that probably didn't belong there.
A note should simply be pitch, velocity, duration.  However, this
implementation has the Note loaded down with it's pattern position
(which is redundant -- the pattern tracks the notes position), ADSR
parameters (this should be managed by the sampler/synth... not on a
per-note basis), and parameters for a low pass resonant filter (again,
this belongs in the sample/synth).  Pan L/R was left in because the
pattern editor supports them.

Some of these parameters could be used to make a really complex type
of sampler/synth... but they're not currently supported by any
editors, song files, or sampler... and they bog down H2Core::Note,
which should be as lightweight as possible.


git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/transport_redesign@871 3b3fb362-3133-0410-aa15-cf69e0a59cb7