Write UTF-8 out instead of TinyXML's lazy thing.
In an attempt to handle non-latin characters, TinyXML was outputting anything
not within 0x20-0x7F as a numeric character reference (e.g. &#x20;).  However,
the UTF-8 sequence D184 was output &#xD1;&#x84; instead of ф or &#x00000444;
(cyrillic small letter EF).  This is technically incorrect, as a numeric
character reference is supposed to be standalone, and not part of an encoding
sequence.

This is corrected by always outputting UTF-8.  Note that &#xD1;&#x84; in a file
will still be interpreted by hydrogen as &#x0444; -- which is technically
incorrect, but makes things backward compatible.  That could use a better
solution.  I.e. if the encoding is specified, these should be read literally.


git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/text_encodings@945 3b3fb362-3133-0410-aa15-cf69e0a59cb7