Add a new_position member to TransportPosition.
This may be premature optimization... but sequencer input objects may want to
keep a "cursor" for keeping track of how far they've progressed in the sequence
(i.e. for lookahead logic).  Instead of making each of them detect relocations,
it made more sense to have it managed by the transport.

git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/transport_redesign_2@1068 3b3fb362-3133-0410-aa15-cf69e0a59cb7