Removed EVENT_STATE/STATE_PLAYING, replaced with EVENT_TRANSPORT.
Hydrogen's STATE now just indicates the health of the audio engine,
and no longer indicates if we are playing.  This is done by checking
the transport.  All of the code is converted to equivalent calls,
and a new EventListener::transportEvent() callback has been added.
However, hydrogen.cpp is still a bit messy since a lot of the
scheduling code was looking at this state.

git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/transport_redesign_2@1082 3b3fb362-3133-0410-aa15-cf69e0a59cb7