Assign the instance pointer early in constructors.
A race condition existed with class H2Core::Hydrogen where the MIDI
driver could call Hydrogen::get_instance() before Hydrogen::__instance
has been assigned.  To fix this, the __instance variable is now
assigned early in the constructor.

While it's likely that this is only an issue with class Hydrogen, it's
a possibility for any singleton class.  Therefore, all constructors
now assign the __instance variable as the first instruction (with a
couple of exceptions).

git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/jack_zombies@1261 3b3fb362-3133-0410-aa15-cf69e0a59cb7