Sampler no longer caches pointer to per-track out buffers.
When starting the audio driver, H2Core::Sampler was keeping a copy of
the output-buffer pointer from JACK and using that to write data.
This is a deprecated practice[1], but appears to be an error whenever
something like the buffer size (frames-per-period) changes.
Therefore, it's believed that this also resolves Ticket #86.[2]

[1] See documentation for jack_port_get_buffer():
    http://jackaudio.org/files/docs/html/group__PortFunctions.html
[2] http://trac.assembla.com/hydrogen/ticket/86

git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/jack_zombies@1271 3b3fb362-3133-0410-aa15-cf69e0a59cb7