Bugfix for #237: jack ports get mixed up when moving instruments. This changesets saves and disconnects all jack-connections before a instrument gets moved and restores all connections afterwards. This solution is not optimal, but a better solution would require a rewrite of our track-output handling.
The biggest problem is that if instrument n gets moved to position n-1,
the jack ports are getting changed as well. This means that jack port n
belongs now to instrument n+1. It would be better if an instruments
would stay with its jack port during its complete lifecycle. A lot of
problem arise by the fact that one single jack port can identify
different instruments in one session (after a move or delete operation).


git-svn-id: http://svn.assembla.com/svn/hydrogen/branches/jackConnectionProblems@2465 3b3fb362-3133-0410-aa15-cf69e0a59cb7